+++ chainer/functions/__init__.py
@@ -36,6 +36,7 @@ from chainer.functions.loss import vae  # NOQA
 from chainer.functions.math import basic_math  # NOQA
 from chainer.functions.math import exponential
 from chainer.functions.math import identity
+from chainer.functions.math import l2_norm_squared
 from chainer.functions.math import matmul
 from chainer.functions.math import minmax
 from chainer.functions.math import sum
@@ -43,7 +44,6 @@ from chainer.functions.math import trigonometric
 from chainer.functions.noise import dropout
 from chainer.functions.noise import gaussian
 from chainer.functions.normalization import batch_normalization
-from chainer.functions.normalization import l2_norm_squared
 from chainer.functions.normalization import local_response_normalization
 from chainer.functions.pooling import average_pooling_2d
 from chainer.functions.pooling import max_pooling_2d
@@ -137,6 +137,8 @@ Exp = exponential.Exp
 exp = exponential.exp
 Identity = identity.Identity
 identity = identity.identity
+L2NormSquared = l2_norm_squared.L2NormSquared
+l2_norm_squared = l2_norm_squared.l2_norm_squared
 Log = exponential.Log
 log = exponential.log
 MatMul = matmul.MatMul
@@ -161,8 +163,6 @@ LocalResponseNormalization = \
     local_response_normalization.LocalResponseNormalization
 local_response_normalization = \
     local_response_normalization.local_response_normalization
-L2NormSquared = l2_norm_squared.L2NormSquared
-l2_norm_squared = l2_norm_squared.l2_norm_squared
 
 AveragePooling2D = average_pooling_2d.AveragePooling2D
 average_pooling_2d = average_pooling_2d.average_pooling_2d
