--- tests/chainer_tests/functions_tests/normalization_tests/test_l2_norm_squared.py
@@ -1,64 +1,0 @@
-import unittest
-
-import numpy
-import six
-
-import chainer
-from chainer import cuda
-from chainer import functions
-from chainer import gradient_check
-from chainer import testing
-from chainer.testing import attr
-from chainer.testing import condition
-
-
-class TestL2NormSquared(unittest.TestCase):
-
-    def setUp(self):
-        self.x = numpy.random.uniform(-1, 1, (2, 7)).astype(numpy.float32)
-        self.gy = numpy.random.uniform(-1, 1, (2, 7)).astype(numpy.float32)
-
-    def check_forward(self, x_data):
-        x = chainer.Variable(x_data)
-        y = functions.l2_norm_squared(x)
-        self.assertEqual(y.data.dtype, numpy.float32)
-        y_data = cuda.to_cpu(y.data)
-
-        y_expect = numpy.zeros(self.x.shape[0])
-        for n in six.moves.range(self.x.shape[0]):
-            y_expect[n] = sum(map(lambda x: x * x, self.x[n]))
-
-        gradient_check.assert_allclose(y_expect, y_data)
-
-    @condition.retry(3)
-    def test_forward_cpu(self):
-        self.check_forward(self.x)
-
-    @attr.gpu
-    @condition.retry(3)
-    def test_forward_gpu(self):
-        self.check_forward(cuda.to_gpu(self.x))
-
-    def check_backward(self, x_data, y_grad):
-        x = chainer.Variable(x_data)
-        y = functions.local_response_normalization(x)
-        y.grad = y_grad
-        y.backward()
-
-        func = y.creator
-        f = lambda: func.forward((x.data,))
-        gx, = gradient_check.numerical_grad(f, (x.data,), (y.grad,), eps=1)
-
-        gradient_check.assert_allclose(gx, x.grad)
-
-    @condition.retry(3)
-    def test_backward_cpu(self):
-        self.check_backward(self.x, self.gy)
-
-    @attr.gpu
-    @condition.retry(3)
-    def test_backward_gpu(self):
-        self.check_backward(cuda.to_gpu(self.x), cuda.to_gpu(self.gy))
-
-
-testing.run_module(__name__, __file__)
